requirements.txt
jira==2.0.0
python_dateutil==2.8.2
